+++ processor/src/test/fixtures/unresolved/com/example/autodetect/provider/AppInjector.java
@@ -1,0 +1,9 @@
+package com.example.autodetect.provider;
+
+import sting.Injector;
+
+@Injector
+interface AppInjector
+{
+  LibModel getLibModel();
+}