+++ processor/src/test/fixtures/unresolved/com/example/autodetect/provider/LibModelImpl.java
@@ -1,0 +1,11 @@
+package com.example.autodetect.provider;
+
+import sting.Injectable;
+import sting.Typed;
+
+@Injectable
+@Typed( LibModel.class )
+public class LibModelImpl
+  extends LibModel
+{
+}